requirements.txt
langchain
langgraph
chromadb
sentence-transformers
openai
pypdf
tiktoken
python-dotenv
typer
langchain-community
langchain-chroma
langchain-openai
langchain_groq
langchain-huggingface